calc geo works now
rho has some weird kinks in that some points are repeated (p0 = p1 = p2 = p3) and it causes a division by 0, which means it returns not real (or complex) outputs, which is not what we need. for now i've made it so that if rho is not a number, that residue is not added to the final residue list.